+++ progress.go
@@ -135,7 +135,7 @@ func (p *Progress) AddTotal(n uint64) {
 //   if total >  0: n represents the relative weight of the work completed, and the progress percentage is calculated as n / totalUnits
 //   if total == 0: n represents the portion of the InitialBudget(), which must be divided among all sub-tasks by the caller
 func (p *Progress) Report(n float64, status string) {
-	p.tracker.store(float64(p.unitsDone.Add(uint64(n))), status)
+	p.tracker.store(p.unitsDone.Add(uint64(n)), status)
 
 	total := p.total.Load()
 
@@ -227,7 +227,7 @@ func (p *Progress) draw(state uint32, val any) {
 
 	if err == nil && p.drawNotify != nil {
 		select {
-		case p.drawNotify <- struct{}{}: // ensures deterministic tests by signaling the completion of a draw cycle
+		case p.drawNotify <- struct{}{}: // ensures synchronous, deterministic tests by signaling the completion of a draw cycle
 		default:
 		}
 	}
@@ -319,7 +319,7 @@ type ticker interface {
 	Stop()
 }
 
-type clock   interface { tick() ticker     } // enables dependency injection to facilitate testing
+type clock   interface { tick() ticker     } // enables dependency injection to facilitate synchronous, deterministic testing
 
 type realTicker struct { *time.Ticker      }
 type fakeTicker struct { c  chan time.Time }
